gsplat_m1_recon/src/utils/config_utils.py
from __future__ import annotations

from pathlib import Path
from typing import Any, Dict

import yaml


def load_config(config_path: str | Path) -> Dict[str, Any]:
    """读取 YAML 配置文件。"""
    path = Path(config_path)
    if not path.exists():
        raise FileNotFoundError(f"配置文件不存在: {path}")
    with path.open("r", encoding="utf-8") as f:
        return yaml.safe_load(f)


def ensure_dirs(paths: list[str | Path]) -> None:
    """批量创建目录（若不存在）。"""
    for p in paths:
        Path(p).mkdir(parents=True, exist_ok=True)
