gsplat_m1_recon/requirements.txt
numpy>=1.24
torch>=2.2
torchvision>=0.17
opencv-python>=4.8
Pillow>=10.0
tqdm>=4.66
matplotlib>=3.8
PyYAML>=6.0
psutil>=5.9
imageio>=2.34
